webapp/osmaxx/requirements/local.txt
# Local development dependencies go here
# -r base.txt is already included in test.txt
-r test.txt
django-extensions
Werkzeug
ipython

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar
